Internalize and threadsafetify String#unicode_normalize
In CRuby, this code was moved into C in 2017 and the .rb file was
removed shortly after (released in Ruby 2.5.0).

This patch does the following:

* Move the script into the JRuby jar so it can be loaded as part of
  our boot logic.
* Fix the race to load UnicodeNormalize by removing the constant
  definition check and restoring the method redefinition (see
  jruby/jruby#9231).

Fixes jruby/jruby#9231 and relates to jruby/jruby#6362.
