+++ core/src/main/ruby/jruby/kernel.rb
@@ -29,3 +29,4 @@ load 'jruby/kernel/method.rb'
 load 'jruby/kernel/thread.rb'
 load 'jruby/kernel/integer.rb'
 load 'jruby/kernel/time.rb'
+load 'jruby/kernel/string.rb'
