--- core/src/main/ruby/jruby/kernel/enc_prelude.rb
@@ -1,4 +1,0 @@
-begin
-  require 'unicode_normalize'
-rescue LoadError
-end
