+++ core/src/main/ruby/jruby/preludes.rb
@@ -1,4 +1,3 @@
 # Load preludes
 load 'jruby/kernel/prelude.rb'
-load 'jruby/kernel/enc_prelude.rb'
 load 'jruby/kernel/gem_prelude.rb'
